# optimize by using generators