Optimize by using generators, so the possible typos don't have to be stored in memory.

For long english words:

%timeit spell('disproporttionatelly')
before:  1 s ± 7.81 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
after:     762 ms ± 18.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit spell('indistimguishabble')
before:   821 ms ± 5.22 ms per loop (mean ± std. dev. of 7 runs, 1 loop each
after:      619 ms ± 10 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)


More importantly, I'm working on a support for polish, where alphabet is larger, and words tend to be longer, so the change is significant:

%timeit spell('gżegrzółka')
before:   1.51 s ± 67.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
after:       370 ms ± 16.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit spell('anarchokolektuwistycznychh')
before:   3.83 s ± 116 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
after:      2.2 s ± 11.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)