@fifimajster 2019-08-31
It also solves issue #18 
Before, running spell('xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx')
would consume all RAM it could. Now it takes only 8KB more than idle, and won't freeze.