@fifimajster 2019-09-05
When using pypy, difference gets even more dramatic:

%timeit spell('disproporttionatelly')
before:  668 ms ± 19.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
after:     377 ms ± 25.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit spell('indistimguishabble')
before:   585 ms ± 15.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
after:      330 ms ± 29 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)


%timeit spell('gżegrzółka')
before:   Gets killed because it eats up too much RAM
before:   166 ms ± 11.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit spell('anarchokolektuwistycznychh')
before:   Gets killed because it eats up too much RAM
after:      994 ms ± 29.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)